Conflicts:
	Sources/Internal/Debug/DVAssertMessageWin.cpp
	Sources/Internal/Network/Base/IfAddress.cpp
	Sources/Internal/Network/Base/NetworkUtils.cpp

+++ Tools/ResourceEditor/Classes/Qt/CubemapEditor/CubeMapTextureBrowser.cpp
@@ -35,6 +35,7 @@
 #include "../../StringConstants.h"
 #include "Scene3D/Systems/SkyboxSystem.h"
 #include "Project/ProjectManager.h"
+#include "ImageTools/ImageTools.h"
 
 #include <QScrollBar>
 
@@ -314,18 +315,21 @@ void CubeMapTextureBrowser::OnDeleteSelectedItemsClicked()
 				FilePath fp = item->data(CUBELIST_DELEGATE_ITEMFULLPATH).toString().toStdString();
 				if(fp.Exists())
 				{
-					DAVA::Vector<DAVA::String> faceNames;
+					DAVA::Vector<DAVA::FilePath> faceNames;
 					CubemapUtils::GenerateFaceNames(fp.GetAbsolutePathname(), faceNames);
 					for(size_t faceIndex = 0; faceIndex < faceNames.size(); ++faceIndex)
 					{
+                        if (faceNames[faceIndex].IsEmpty())
+                            continue;
+
 						FilePath hackTex = faceNames[faceIndex];
 						hackTex.ReplaceExtension(".tex");
 						
 						QFile::remove(hackTex.GetAbsolutePathname().c_str());
-						bool removeResult = QFile::remove(faceNames[faceIndex].c_str());
+						bool removeResult = QFile::remove(faceNames[faceIndex].GetAbsolutePathname().c_str());
 						if(!removeResult)
 						{
-							failedToRemove.push_back(faceNames[faceIndex]);
+                            failedToRemove.push_back(faceNames[faceIndex].GetAbsolutePathname().c_str());
 						}
 					}
 					
@@ -368,12 +372,15 @@ bool CubeMapTextureBrowser::ValidateTextureAndFillThumbnails(DAVA::FilePath& fp,
 	
 	int width = 0;
 	int height = 0;
-	DAVA::Vector<DAVA::String> faceNames;
+	DAVA::Vector<DAVA::FilePath> faceNames;
 	CubemapUtils::GenerateFaceNames(fp.GetAbsolutePathname(), faceNames);
 	for(size_t i = 0; i < faceNames.size(); ++i)
 	{
-		QImage faceImage;
-		if(!faceImage.load(faceNames[i].c_str())) //file must be present
+        if (faceNames[i].IsEmpty())
+            continue;
+
+		QImage faceImage = ImageTools::FromDavaImage(faceNames[i]);
+		if(faceImage.isNull()) //file must be present
 		{
 			result = false;
 		}
